# Conflicts:
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/Components/RoutineInfo.kt

+++ fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/data/network/model/cycleExercises/NetworkCycleExercise.kt
@@ -9,7 +9,7 @@ data class NetworkCycleExercise (
     @SerializedName("name"   ) var name   : String,
     @SerializedName("detail" ) var detail : String,
     @SerializedName("type"   ) var type   : String,
-    @SerializedName("date"   ) var date   : Int,
+    @SerializedName("date"   ) var date   : Long,
     @SerializedName("order"  ) var order  : Int
 
 )
